@@ -4,11 +4,13 @@
   "description": "kanban board using just GitHub",
   "main": "index.js",
   "scripts": {
-    "lint": "eslint src --ext .js --ext .jsx",
-    "start": "webpack-dev-server --progress --colors --display-chunks --display-modules --display-origins",
-    "build": "npm run lint; NODE_ENV=production webpack --optimize-minimize --progress --colors",
+    "lint": "eslint src test --ext .js --ext .jsx",
+    "start": "webpack-dev-server --host 0.0.0.0 --port 8080 --hot --progress",
+    "serve": "http-server .",
+    "build-and-test": "./script/run-build-and-then-test.sh",
+    "build": "npm run lint; NODE_ENV=production webpack --progress --colors",
     "dev-build": "npm run lint; webpack --progress --colors",
-    "start-webdriver": "webdriver-manager update && webdriver-manager start",
+    "start-selenium": "selenium-standalone install && selenium-standalone start",
     "test:only": "ava --verbose --serial",
     "test": "./script/run-test.sh",
     "deploy": "./script/deploy.sh"
@@ -16,32 +18,35 @@
   "author": "Philip Schatz",
   "license": "AGPL3.0",
   "devDependencies": {
-    "ava": "^0.18.0",
+    "ava": "^0.25.0",
     "babel-core": "^6.3.17",
     "babel-eslint": "^7.1.0",
     "babel-loader": "^7.0.0",
-    "babel-plugin-transform-object-rest-spread": "^6.20.2",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "css-loader": "^0.28.1",
+    "babel-plugin-react-require": "^3.0.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "css-loader": "^0.28.7",
     "eslint": "^3.8.1",
     "eslint-loader": "^1.1.1",
     "eslint-plugin-react": "^7.0.0",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^3.0.0",
     "favico.js": "^0.3.10",
     "file-loader": "^0.11.0",
-    "http-server": "^0.9.0",
+    "http-server": "^0.11.1",
     "json-loader": "^0.5.4",
-    "less": "^2.5.3",
+    "less": "^2.7.1",
     "less-loader": "^4.0.4",
     "node-libs-browser": "^2.0.0",
-    "parallelshell": "^2.0.0",
-    "react-hot-loader": "^3.0.0-beta.6",
+    "npm-run-all": "^4.1.3",
+    "selenium-standalone": "^6.15.1",
+    "style-loader": "^0.18.2",
+    "uglifyjs-webpack-plugin": "^1.0.1",
     "url-loader": "^0.5.6",
-    "webdriver-manager": "^11.1.0",
-    "webdriverio": "^4.5.2",
-    "webpack": "^1.9.10",
-    "webpack-dev-server": "^1.9.0"
+    "webdriverio": "^4.12.0",
+    "webpack": "^3.12.0",
+    "webpack-dev-server": "^2.11.2"
   },
   "dependencies": {
     "babel-polyfill": "^6.3.14",
@@ -64,13 +69,14 @@
     "octicons": "^6.0.0",
     "octokat": "^0.7.0",
     "puzzle-script": "0.0.2",
+    "prop-types": "^15.6.1",
     "qs": "^6.3.0",
-    "react": "^15.3.2",
-    "react-bootstrap": "^0.31.0",
+    "react": "^16.3.2",
+    "react-bootstrap": "^0.31.5",
     "react-dnd": "^2.1.2",
     "react-dnd-html5-backend": "^2.1.2",
-    "react-dom": "^15.3.2",
-    "react-octicons": "0.0.5",
+    "react-dom": "^16.3.2",
+    "react-octicons": "0.2.0",
     "react-router": "^3.0.0",
     "recordo": "^0.0.6",
     "ultramarked": "^1.7.0",
