@@ -1,4 +1,4 @@
-import React from 'react';
+import {Component} from 'react';
 import HTML5Backend from 'react-dnd-html5-backend';
 import { DragDropContext } from 'react-dnd';
 
@@ -9,33 +9,39 @@ import history from '../../history';
 import AppNav from './nav';
 import Footer from './footer';
 
-const App = React.createClass({
+class App extends Component {
   componentDidMount() {
     SettingsStore.on('change:tableLayout', this.onChange);
     this._historyListener = history.listen(this.storeHistory);
-  },
-  componentWillMount() {
+  }
+
+  UNSAFE_componentWillMount() {
     SettingsStore.off('change:tableLayout', this.onChange);
     setRouterHistory(history);
     setFilters(this.props);
-  },
-  componentWillUpdate(newProps) {
+  }
+
+  UNSAFE_componentWillUpdate(newProps) {
     setFilters(newProps);
-  },
+  }
+
   componentWillUnmount() {
     this._historyListener();
-  },
-  storeHistory(locationChangeEvent) {
+  }
+
+  storeHistory = () => {
     if (window.ga) {
       const {pathname, hash} = window.location;
       window.ga('set', 'page', pathname + hash);
       // window.ga('set', 'page', '/gh-board/#' + locationChangeEvent.pathname + locationChangeEvent.search);
       window.ga('send', 'pageview');
     }
-  },
-  onChange() {
+  };
+
+  onChange = () => {
     this.forceUpdate();
-  },
+  };
+
   render() {
     const {params} = this.props;
     const classes = ['app'];
@@ -49,6 +55,6 @@ const App = React.createClass({
       </div>
     );
   }
-});
+}
 
 export default DragDropContext(HTML5Backend)(App);
