@@ -1,6 +1,5 @@
-import React from 'react';
+import {Component} from 'react';
 import _ from 'underscore';
-import * as BS from 'react-bootstrap';
 import {Link} from 'react-router';
 import {MilestoneIcon} from 'react-octicons';
 
@@ -16,127 +15,123 @@ import Board from './board';
 import GithubFlavoredMarkdown from './gfm';
 
 
-const KanbanColumn = React.createClass({
-  render() {
-    const {milestone, cards, primaryRepoName, columnRegExp} = this.props;
+function KanbanColumn(props) {
+  const {milestone, cards, primaryRepoName, columnRegExp} = props;
 
-    const issueComponents = _.map(cards, (card) => {
-      return (
-        <Issue
-          key={card.key()}
-          primaryRepoName={primaryRepoName}
-          card={card}
-          columnRegExp={columnRegExp}
-          />
-      );
-    });
-
-    let heading;
-    if (milestone) {
-      heading = (
-        <Link className='milestone-title' to={getFilters().toggleMilestoneTitle(milestone.title).url()}>
-          <MilestoneIcon/>
-          <GithubFlavoredMarkdown
-            inline
-            disableLinks={true}
-            text={milestone.title}/>
-        </Link>
-      );
-    } else {
-      heading = 'No Milestone';
-    }
-
-    const isShowingColumn = (
-         (!milestone && !SettingsStore.getHideUncategorized())
-      || (milestone && (
-             getFilters().getState().milestoneTitles.length == 0
-          || getFilters().getState().milestoneTitles.indexOf(milestone.title) >= 0))
+  const issueComponents = _.map(cards, (card) => {
+    return (
+      <Issue
+        key={card.key()}
+        primaryRepoName={primaryRepoName}
+        card={card}
+        columnRegExp={columnRegExp}
+        />
     );
+  });
 
-    if (isShowingColumn) {
-      return (
-        <div key={milestone && milestone.title || 'no-milestone'} className='kanban-board-column'>
-          <IssueList
-            title={heading}
-            milestone={milestone}
-          >
-            {issueComponents}
-          </IssueList>
-        </div>
-      );
-    } else {
-      return null; // TODO: Maybe the panel should say "No Issues" (but only if it's the only column)
-    }
-
+  let heading;
+  if (milestone) {
+    heading = (
+      <Link className='milestone-title' to={getFilters().toggleMilestoneTitle(milestone.title).url()}>
+        <MilestoneIcon/>
+        <GithubFlavoredMarkdown
+          inline
+          disableLinks={true}
+          text={milestone.title}/>
+      </Link>
+    );
+  } else {
+    heading = 'No Milestone';
   }
-});
 
+  const isShowingColumn = (
+       (!milestone && !SettingsStore.getHideUncategorized())
+    || (milestone && (
+           getFilters().getState().milestoneTitles.length == 0
+        || getFilters().getState().milestoneTitles.indexOf(milestone.title) >= 0))
+  );
 
-const ByMilestoneView = React.createClass({
-  render() {
-    const {columnData, cards, primaryRepoName, columnRegExp} = this.props;
+  if (isShowingColumn) {
+    return (
+      <div key={milestone && milestone.title || 'no-milestone'} className='kanban-board-column'>
+        <IssueList
+          title={heading}
+          milestone={milestone}
+        >
+          {issueComponents}
+        </IssueList>
+      </div>
+    );
+  } else {
+    return null; // TODO: Maybe the panel should say "No Issues" (but only if it's the only column)
+  }
 
-    const uncategorizedCards = _.filter(cards, (card) => {
-      return !card.issue.milestone;
+}
+
+function ByMilestoneView(props) {
+  const {columnData, cards, primaryRepoName, columnRegExp} = props;
+
+  const uncategorizedCards = _.filter(cards, (card) => {
+    return !card.issue.milestone;
+  });
+
+  let sortedCards = FilterStore.filterAndSort(cards, true/*isShowingMilestones*/);
+
+  let kanbanColumnCount = 0; // Count the number of actual columns displayed
+  kanbanColumnCount++; // "No Milestone" counts as a column
+
+  const uncategorizedColumn = (
+    <KanbanColumn
+      cards={uncategorizedCards}
+      primaryRepoName={primaryRepoName}
+      columnRegExp={columnRegExp}
+    />
+  );
+
+  const kanbanColumns = _.map(columnData, (milestone) => {
+    // If we are filtering by a kanban column then only show that column
+    // Otherwise show all columns
+    const columnCards = _.filter(sortedCards, (card) => {
+      return card.issue.milestone && card.issue.milestone.title === milestone.title;
     });
 
-    let sortedCards = FilterStore.filterAndSort(cards, true/*isShowingMilestones*/);
 
-    let kanbanColumnCount = 0; // Count the number of actual columns displayed
-    kanbanColumnCount++; // "No Milestone" counts as a column
+    // Show the column when:
+    // isFilteringByColumn = label (the current column we are filtering on)
+    // !isFilteringByColumn && (!getShowEmptyColumns || columnCards.length)
 
-    const uncategorizedColumn = (
+    kanbanColumnCount++; // Count the number of actual columns displayed
+    /*HACK: Column should handle milestones */
+    return (
       <KanbanColumn
-        cards={uncategorizedCards}
+        key={kanbanColumnCount}
+        milestone={milestone}
+        cards={columnCards}
         primaryRepoName={primaryRepoName}
         columnRegExp={columnRegExp}
       />
     );
+  });
 
-    const kanbanColumns = _.map(columnData, (milestone) => {
-      // If we are filtering by a kanban column then only show that column
-      // Otherwise show all columns
-      const columnCards = _.filter(sortedCards, (card) => {
-        return card.issue.milestone && card.issue.milestone.title === milestone.title;
-      });
+  return (
+    <div className='kanban-board'>
+      {uncategorizedColumn}
+      {kanbanColumns}
+    </div>
+  );
+}
 
-
-      // Show the column when:
-      // isFilteringByColumn = label (the current column we are filtering on)
-      // !isFilteringByColumn && (!getShowEmptyColumns || columnCards.length)
-
-      kanbanColumnCount++; // Count the number of actual columns displayed
-      /*HACK: Column should handle milestones */
-      return (
-        <KanbanColumn
-          key={kanbanColumnCount}
-          milestone={milestone}
-          cards={columnCards}
-          primaryRepoName={primaryRepoName}
-          columnRegExp={columnRegExp}
-        />
-      );
-    });
-
-    return (
-      <div className='kanban-board'>
-        {uncategorizedColumn}
-        {kanbanColumns}
-      </div>
-    );
-  }
-});
-
-
-const RepoKanbanShell = React.createClass({
-  componentWillMount() {
+class RepoKanbanShell extends Component {
+  UNSAFE_componentWillMount() {
     // Needs to be called before `render()`
     IssueStore.startPolling();
-  },
+  }
+
   componentWillUnmount() {
     IssueStore.stopPolling();
-  },
-  renderLoaded() {
+  }
+
+  renderLoaded = () => {
     const {repoInfos, columnRegExp} = getFilters().getState();
     // pull out the primaryRepoName
     const [{repoOwner, repoName}] = repoInfos;
@@ -149,7 +144,8 @@ const RepoKanbanShell = React.createClass({
         columnDataPromise={IssueStore.fetchMilestones(repoOwner, repoName)}
       />
     );
-  },
+  };
+
   render() {
 
     return (
@@ -159,6 +155,6 @@ const RepoKanbanShell = React.createClass({
       />
     );
   }
-});
+}
 
 export default RepoKanbanShell;
