@@ -1,12 +1,11 @@
-import React from 'react';
+import {Component} from 'react';
 
 import Client from '../github-client';
-import IssueStore from '../issue-store';
 
 import Etherpad from './etherpad';
 
 
-const EtherpadIssueShell = React.createClass({
+class EtherpadIssueShell extends Component {
   render() {
     const {repoOwner, repoName, branch, splat} = this.props.params;
     const path = splat;
@@ -26,7 +25,7 @@ const EtherpadIssueShell = React.createClass({
       });
     };
     const loadBody = () => {
-      return Client.getOcto().repos(repoOwner, repoName).contents(path).fetch({ref: branch}).then(({content, sha, encoding}) => {
+      return Client.getOcto().repos(repoOwner, repoName).contents(path).fetch({ref: branch}).then(({content, sha}) => {
         fileText = atob(content);
         fileSha = sha;
         return fileText;
@@ -38,6 +37,6 @@ const EtherpadIssueShell = React.createClass({
       <Etherpad promise={promise} title={title} padName={padName} getBody={getBody} saveBody={saveBody} loadBody={loadBody} repoOwner={repoOwner} repoName={repoName}/>
     );
   }
-});
+}
 
 export default EtherpadIssueShell;
