@@ -251,7 +251,12 @@ const database = new class Database {
   }
   putCards(cards) {
     const batchOps = cards.map((card) => {
-      const {repoOwner, repoName, number} = card;
+      let {repoOwner, repoName, number} = card;
+
+      if (!number) {
+        number = card.issue.number;
+      }
+
       const value = this.toCardValue(card);
       return {
         type: 'put',
