@@ -1,14 +1,11 @@
 const path = require('path');
 const webpack = require('webpack');
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 
 const isBuild = process.env['NODE_ENV'] === 'production';
 
-const config = {
-  // devtool: '#eval-source-map',
-  debug: !isBuild,
-  devtool: 'source-map',
-  context: path.resolve(__dirname),
+module.exports = {
   entry: [
     './style/index.js',
     './src/index.js'
@@ -20,45 +17,63 @@ const config = {
     publicPath: isBuild ? './dist/' : '/dist/', // gh-pages needs this to have a '.' for bundles
     filename: 'bundle.js'
   },
-  plugins: [
-    new webpack.NoErrorsPlugin(),
-    new ExtractTextPlugin('app.css')
-  ].concat(isBuild ? [
-    new webpack.DefinePlugin({ 'process.env': { NODE_ENV: JSON.stringify('production') } })
-  ] : []),
+  context: path.resolve(__dirname),
+  devtool: isBuild ? false : 'source-map',
   module: {
-    preLoaders: [
-      { test: /\.jsx?$/, loader: 'eslint-loader', exclude: [/node_modules|gantt-chart.*/, /octokat\.js/] },
-    ],
-    loaders: [
+    rules: [
       {
         test: /\.jsx?$/,
-        loader: 'babel',
         exclude: [/node_modules/, /puzzle-script/, /octokat\.js/],
-        query: {
-          presets: ['react', 'es2015'],
-          plugins: ['transform-object-rest-spread']
-        }
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: [
+              'react',
+              'env'
+            ],
+            plugins: [
+              'react-require',
+              'transform-object-rest-spread',
+              'transform-class-properties'
+            ],
+          }
+        },
       },
-      { test: /\.json$/, loader: 'json-loader'},
-      { test: /\.less$/,  loader: ExtractTextPlugin.extract('css!less') },
-      { test: /\.(png|jpg|svg)/, loader: 'file-loader?name=[name].[ext]'},
-      { test: /\.(woff|woff2|eot|ttf)/, loader: 'url-loader?limit=30000&name=[name]-[hash].[ext]' }
+      {
+        test: /\.less$/,
+        use: isBuild ? ExtractTextPlugin.extract({
+          fallback: 'style-loader',
+          use: [
+            {
+              loader: 'css-loader',
+              options: {
+                minimize: true
+              }
+            },
+            'less-loader'
+          ]
+        }) :
+        ['style-loader', 'css-loader', 'less-loader']
+      },
+      { test: /\.(png|jpg|svg)/, use: 'file-loader?name=[name].[ext]'},
+      { test: /\.(woff|woff2|eot|ttf)/, use: 'url-loader?limit=30000&name=[name]-[hash].[ext]' }
     ]
   },
   resolve: {
-    extensions: ['', '.js', '.jsx', '.json'],
+    extensions: ['.js', '.jsx', '.json'],
     alias: {
       xmlhttprequest: path.join(__dirname, '/src/hacks/xmlhttprequest-filler.js'),
       fs: path.join(__dirname, '/src/hacks/mermaid-stubs.js'),
       proxyquire: path.join(__dirname, '/src/hacks/mermaid-stubs.js'),
       rewire: path.join(__dirname, '/src/hacks/mermaid-stubs.js'),
       'mock-browser': path.join(__dirname, '/src/hacks/mermaid-stubs.js')
-    },
+    }
   },
-  devServer: {
-    hotComponents: !isBuild
-  }
+  plugins: isBuild ? [
+    new webpack.DefinePlugin({
+      'process.env.NODE_ENV': JSON.stringify('production')
+    }),
+    new ExtractTextPlugin('app.css'),
+    new UglifyJsPlugin()
+  ] : []
 };
-
-module.exports = config;
